backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
asyncpg==0.30.0
redis==5.2.1
openai==1.59.6
azure-ai-vision-imageanalysis==1.0.0
azure-storage-blob==12.24.0
pillow==11.1.0
python-dotenv==1.0.1
aiofiles==24.1.0
aiohttp==3.11.11
python-jose==3.3.0
httpx==0.28.1
